requirements.txt
filelock==3.14.0
fsspec==2024.6.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pillow==10.3.0
sympy==1.12.1
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.1
